requirements.txt
Flask
Flask-SQLAlchemy
sqlalchemy
flask-login
flask-wtf 
email_validator
requests
libsass
flask-assets 
Flask-Migrate
python-dotenv
